pyproject.toml
[tool.poetry]
name = "weather-data"
version = "0.1.0"
description = ""
authors = ["igornevertired <igornevertired@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "3.1.5"
sqlalchemy = "2.0.35"
requests = "2.32.3"
pyyaml = "6.0.2"
psycopg2 = "2.9.9"
sqlalchemy-utils = "0.41.2"
numpy = "2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
